src/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap4==2024.10
debugpy==1.8.1
decorator==5.1.1
Django==5.1.1
django-allauth==65.0.2
django-bootstrap4==24.3
django-crispy-forms==2.3
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.8
pillow==10.4.0
pycountry==24.6.1
python-slugify==8.0.4
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0
